requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
beautifulsoup4==4.12.2
requests==2.31.0
httpx[http2]==0.25.2
# Playwright for browser automation (replaces Selenium)
playwright==1.40.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
APScheduler==3.10.4
lxml>=5.0.0
fake-useragent==1.4.0
# FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6
# Celery + Redis dependencies
celery==5.3.4
redis>=5.0.1

